pyproject.toml
[project]
name = "txsoundgen"
version = "0.1.0"
description = "Sound pack generator for EdgeTX/OpenTX radios"
license = "MIT"
readme = "README.md"
authors = [
    { name = "Ben Fairless", email = "ben.fairless@gmail.com" }
]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: Communications :: Ham Radio",
    "Topic :: Multimedia :: Sound/Audio :: Speech"
]
requires-python = ">=3.13"
dependencies = [
    "boto3 (>=1.40.41,<2.0.0)",
    "peewee (>=3.18.2,<4.0.0)",
    "coloredlogs (>=15.0.1,<16.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "pdoc (>=15.0.4,<16.0.0)",
    "pylint (>=3.3.8,<4.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "pytest-spec (>=5.1.0,<6.0.0)",
    "python-magic (>=0.4.27,<0.5.0)",
    "black (>=25.9.0,<26.0.0)",
    "invoke (>=2.2.0,<3.0.0)"

]

[project.urls]
source = "https://github.com/benfairless/txsoundgen"
issues = "https://github.com/benfairless/txsoundgen/issues"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
spec_header_format = "[{module_path}] {test_case}:"
spec_test_format = "{result} {name} : {docstring_summary}"
addopts = "--spec"

[tool.pylint.format]
# max-line-length = 88

[tool.pylint.typecheck]
ignored-classes = ["Wave_read"]
